Optimize edge case of `ln()`.
Swapped the division and the multiplication of the imaginary part. This can potentially execute faster, because the multiplication is likely finished before the divisor becomes available to the CPU instruction pipeline. Previously, the CPU would have to wait for the divisor to become available and the division would then be followed by another multiplication. This would prevent out of order execution and increase the latency of the computation.